refactor: consolidate loggers into single mail/utils/logger.py module
Now that all four loggers share EventLogger, collapse the
mail/utils/logger/ package (base/push/storage/outbound/inbound) into one
mail/utils/logger.py module holding the base class, the four subclasses,
and the get_*_logger factories. Update the five import sites accordingly.

(cherry picked from commit b060410978f6dc547ab175bbc7aa3ce657e997a2)
