+++ mail/utils/logger.py
@@ -3,7 +3,7 @@ from typing import Any
 import frappe
 from frappe.utils import cint
 
-from mail.utils import get_config
+from mail.utils import get_mail_config
 
 
 class EventLogger:
@@ -27,7 +27,7 @@ class EventLogger:
 
 	Subclasses set `logger_name` (the frappe logger channel, e.g. "mail.push")
 	and `config_prefix` (the Mail Settings key prefix, e.g. "push"), which
-	selects the `<prefix>_log_max_file_size`, `<prefix>_log_file_count` and
+	selects the `<prefix>_log_max_size`, `<prefix>_log_file_count` and
 	`<prefix>_log_level` config values.
 	"""
 
@@ -35,9 +35,9 @@ class EventLogger:
 	config_prefix: str
 
 	def __init__(self, ctx: dict | None = None) -> None:
-		config = get_config()
+		config = get_mail_config()
 
-		max_size = cint(config[f"{self.config_prefix}_log_max_file_size"])
+		max_size = cint(config[f"{self.config_prefix}_log_max_size"])
 		file_count = cint(config[f"{self.config_prefix}_log_file_count"])
 		self.logger = frappe.logger(
 			self.logger_name, allow_site=True, max_size=max_size, file_count=file_count
@@ -63,3 +63,71 @@ class EventLogger:
 
 	def exception(self, event: str, **fields: Any) -> None:
 		self.logger.exception(self._record(event, fields))
+
+
+class PushLogger(EventLogger):
+	"""Structured event logger for mail push notifications ("mail.push")."""
+
+	logger_name = "mail.push"
+	config_prefix = "push"
+
+
+class StorageLogger(EventLogger):
+	"""Structured event logger for mail storage operations ("mail.storage")."""
+
+	logger_name = "mail.storage"
+	config_prefix = "storage"
+
+
+class OutboundLogger(EventLogger):
+	"""Structured event logger for outbound mail operations ("mail.outbound")."""
+
+	logger_name = "mail.outbound"
+	config_prefix = "outbound"
+
+
+class InboundLogger(EventLogger):
+	"""Structured event logger for inbound mail operations ("mail.inbound")."""
+
+	logger_name = "mail.inbound"
+	config_prefix = "inbound"
+
+
+def get_push_logger(ctx: dict | None = None) -> PushLogger:
+	"""Returns a structured event logger for mail push notifications.
+
+	The returned logger is bound to `ctx` (by reference); mutating that same
+	dict between log calls is reflected in subsequent records.
+	"""
+
+	return PushLogger(ctx)
+
+
+def get_storage_logger(ctx: dict | None = None) -> StorageLogger:
+	"""Returns a structured event logger for mail storage operations.
+
+	The returned logger is bound to `ctx` (by reference); mutating that same
+	dict between log calls is reflected in subsequent records.
+	"""
+
+	return StorageLogger(ctx)
+
+
+def get_outbound_logger(ctx: dict | None = None) -> OutboundLogger:
+	"""Returns a structured event logger for outbound mail operations.
+
+	The returned logger is bound to `ctx` (by reference); mutating that same
+	dict between log calls is reflected in subsequent records.
+	"""
+
+	return OutboundLogger(ctx)
+
+
+def get_inbound_logger(ctx: dict | None = None) -> InboundLogger:
+	"""Returns a structured event logger for inbound mail operations.
+
+	The returned logger is bound to `ctx` (by reference); mutating that same
+	dict between log calls is reflected in subsequent records.
+	"""
+
+	return InboundLogger(ctx)
